// src/interfaces/approach.interface.ts
/**
 * ### Подход.
 * #### Включает в себя:
 * - <b>approachNumber: number</b> > порядковый номер подхода
 * - <b>numberOfReps: number</b> > кол-во повторений
 * - <b>weight: number</b> > вес
 */
export interface IApproach {
    approachNumber?: number;
    numberOfReps?: number;
    weight?: string;
}